+++ src/ultrasphere/harmonics/expansion.py
@@ -6,16 +6,17 @@ from array_api._2024_12 import Array
 from array_api_compat import array_namespace
 
 from ultrasphere.coordinates import (
-    BranchingType,
     SphericalCoordinates,
     TEuclidean,
     TSpherical,
-    get_child,
 )
-from ultrasphere.harmonics.assume import get_n_end_and_include_negative_m_from_expansion
+from ultrasphere.harmonics.assume import (
+    get_n_end_and_include_negative_m_from_expansion,
+    ndim_harmonics,
+)
 from ultrasphere.integral import integrate
 
-from .expansion import ndim_harmonics as ndim_harmonics_
+from .assume import ndim_harmonics as ndim_harmonics_
 from .harmonics import harmonics as harmonics_
 from .harmonics import harmonics as harmonics__
 
@@ -353,162 +354,3 @@ def expand(
         device=device,
         dtype=dtype,
     )
-
-
-def expand_dim_harmoncis(
-    c: SphericalCoordinates[TSpherical, TEuclidean],
-    node: TSpherical,
-    harmonics: Array,
-) -> Array:
-    """
-    Expand the dimension of the harmonics.
-
-    Expand the dimension so that
-    all values of the harmonics() result dictionary
-    are commomly indexed by the same s_nodes, by default False
-
-    For example, if spherical coordinates,
-    if True, the result will be indexed {"phi": [m], "theta": [m, n]}
-    if False, the result will be indexed {"phi": [m, n], "theta": [m, n]}
-
-    Note that the values will not be repeated
-    therefore the computational cost will be the same
-
-    Parameters
-    ----------
-    node : TSpherical
-        The node of the spherical coordinates.
-    harmonics : Array
-        The harmonics (eigenfunctions).
-
-    Returns
-    -------
-    Array
-        The expanded harmonics.
-        The shapes does not need to be either
-        same or broadcastable.
-
-    """
-    xp = array_namespace(harmonics)
-    idx_node = c.s_nodes.index(node)
-    branching_type = c.branching_types[node]
-    if branching_type == BranchingType.A:
-        moveaxis = {0: idx_node}
-    elif branching_type == BranchingType.B:
-        idx_sin_child = c.s_nodes.index(get_child(c.G, node, "sin"))
-        moveaxis = {
-            0: idx_sin_child,
-            1: idx_node,
-        }
-    elif branching_type == BranchingType.BP:
-        idx_cos_child = c.s_nodes.index(get_child(c.G, node, "cos"))
-        moveaxis = {
-            0: idx_cos_child,
-            1: idx_node,
-        }
-    elif branching_type == BranchingType.C:
-        idx_cos_child = c.s_nodes.index(get_child(c.G, node, "cos"))
-        idx_sin_child = c.s_nodes.index(get_child(c.G, node, "sin"))
-        moveaxis = {0: idx_cos_child, 1: idx_sin_child, 2: idx_node}
-    value_additional_ndim = harmonics.ndim - len(moveaxis)
-    moveaxis = {
-        k + value_additional_ndim: v + value_additional_ndim
-        for k, v in moveaxis.items()
-    }
-    adding_ndim = c.s_ndim - len(moveaxis)
-    harmonics = harmonics[(...,) + (None,) * adding_ndim]
-    return xp.moveaxis(harmonics, list(moveaxis.keys()), list(moveaxis.values()))
-
-
-def expand_dims_harmonics(
-    c: SphericalCoordinates[TSpherical, TEuclidean],
-    harmonics: Mapping[TSpherical, Array],
-) -> Mapping[TSpherical, Array]:
-    """
-    Expand dimensions of the harmonics.
-
-    Expand dimensions so that
-    all values of the harmonics() result dictionary
-    are commomly indexed by the same s_nodes, by default False
-
-    For example, if spherical coordinates,
-    if True, the result will be indexed {"phi": [m], "theta": [m, n]}
-    if False, the result will be indexed {"phi": [m, n], "theta": [m, n]}
-
-    Note that the values will not be repeated
-    therefore the computational cost will be the same
-
-    Parameters
-    ----------
-    harmonics : Mapping[TSpherical, Array]
-        The dictionary of harmonics (eigenfunctions).
-
-    Returns
-    -------
-    Mapping[TSpherical, Array]
-        The expanded harmonics.
-        The shapes does not need to be either
-        same or broadcastable.
-
-    """
-    result: dict[TSpherical, Array] = {}
-    for node in c.s_nodes:
-        result[node] = expand_dim_harmoncis(c, node, harmonics[node])
-    return result
-
-
-def concat_harmonics(
-    c: SphericalCoordinates[TSpherical, TEuclidean],
-    harmonics: Mapping[TSpherical, Array],
-) -> Array:
-    """
-    Concatenate the mapping of expanded harmonics.
-
-    Parameters
-    ----------
-    harmonics : Mapping[TSpherical, Array]
-        The expanded harmonics.
-
-    Returns
-    -------
-    Array
-        The concatenated harmonics.
-
-    """
-    xp = array_namespace(*harmonics.values())
-    try:
-        if c.s_ndim == 0:
-            return xp.asarray(1)
-        return xp.prod(
-            xp.stack(xp.broadcast_arrays(*harmonics.values()), axis=0), axis=0
-        )
-    except Exception as e:
-        shapes = {k: v.shape for k, v in harmonics.items()}
-        raise RuntimeError(f"Error occurred while concatenating {shapes=}") from e
-
-
-def ndim_harmonics(
-    c: SphericalCoordinates[TSpherical, TEuclidean],
-    node: TSpherical,
-) -> int:
-    """
-    The number of dimensions of the eigenfunction
-    corresponding to the node.
-
-    Parameters
-    ----------
-    node : TSpherical
-        The node of the spherical coordinates.
-
-    Returns
-    -------
-    int
-        The number of dimensions.
-
-    """
-    return {
-        BranchingType.A: 1,
-        BranchingType.B: 2,
-        BranchingType.BP: 2,
-        BranchingType.C: 3,
-    }[c.branching_types[node]]
