+++ tests/test_cooridnates.py
@@ -15,8 +15,9 @@ from ultrasphere.creation import (
     random,
     standard,
 )
+from ultrasphere.harmonics.assume import ndim_harmonics
 from ultrasphere.harmonics.cut import expand_cut
-from ultrasphere.harmonics.expansion import expand, expand_evaluate, ndim_harmonics
+from ultrasphere.harmonics.expansion import expand, expand_evaluate
 from ultrasphere.harmonics.flatten import (
     flatten_harmonics,
     flatten_mask_harmonics,
